+++ packages/client/src/StreamSfuClient.ts
@@ -25,15 +25,16 @@ import {
 } from './gen/video/sfu/signal_rpc/signal';
 import { ICETrickle, TrackType } from './gen/video/sfu/models/models';
 import { StreamClient } from './coordinator/connection/client';
-import { generateUUIDv4, sleep } from './coordinator/connection/utils';
+import { generateUUIDv4 } from './coordinator/connection/utils';
 import { Credentials } from './gen/coordinator';
 import { Logger } from './coordinator/connection/types';
 import { getLogger, getLogLevel } from './logger';
-import { withoutConcurrency } from './helpers/concurrency';
 import {
   promiseWithResolvers,
   PromiseWithResolvers,
-} from './helpers/withResolvers';
+  makeSafePromise,
+  SafePromise,
+} from './helpers/promise';
 
 export type StreamSfuClientConstructor = {
   /**
@@ -101,7 +102,7 @@ export class StreamSfuClient {
   /**
    * Promise that resolves when the WebSocket connection is ready (open).
    */
-  private signalReady!: Promise<WebSocket>;
+  private signalReady!: SafePromise<WebSocket>;
 
   /**
    * Flag to indicate if the client is in the process of leaving the call.
@@ -116,7 +117,6 @@ export class StreamSfuClient {
   private pingIntervalInMs = 10 * 1000;
   private unhealthyTimeoutInMs = this.pingIntervalInMs + 5 * 1000;
   private lastMessageTimestamp?: Date;
-  private readonly restoreWebSocketConcurrencyTag = Symbol('recoverWebSocket');
   private readonly unsubscribeIceTrickle: () => void;
   private readonly unsubscribeNetworkChanged: () => void;
   private readonly onSignalClose: (() => void) | undefined;
@@ -124,7 +124,7 @@ export class StreamSfuClient {
   private readonly logTag: string;
   private readonly credentials: Credentials;
   private readonly dispatcher: Dispatcher;
-  private readonly joinResponseTimeout?: number;
+  private readonly joinResponseTimeout: number;
   private networkAvailableTask: PromiseWithResolvers<void> | undefined;
   /**
    * Promise that resolves when the JoinResponse is received.
@@ -227,30 +227,29 @@ export class StreamSfuClient {
     });
 
     this.signalWs.addEventListener('close', this.handleWebSocketClose);
-    this.signalWs.addEventListener('error', this.restoreWebSocket);
 
-    this.signalReady = new Promise((resolve) => {
-      const onOpen = () => {
-        this.signalWs.removeEventListener('open', onOpen);
-        resolve(this.signalWs);
-      };
-      this.signalWs.addEventListener('open', onOpen);
-    });
+    this.signalReady = makeSafePromise(
+      Promise.race<WebSocket>([
+        new Promise((resolve) => {
+          const onOpen = () => {
+            this.signalWs.removeEventListener('open', onOpen);
+            resolve(this.signalWs);
+          };
+          this.signalWs.addEventListener('open', onOpen);
+        }),
+
+        new Promise((resolve, reject) => {
+          setTimeout(
+            () => reject(new Error('SFU WS connection timed out')),
+            this.joinResponseTimeout,
+          );
+        }),
+      ]),
+    );
   };
 
   private cleanUpWebSocket = () => {
-    this.signalWs.removeEventListener('error', this.restoreWebSocket);
     this.signalWs.removeEventListener('close', this.handleWebSocketClose);
-  };
-
-  private restoreWebSocket = () => {
-    withoutConcurrency(this.restoreWebSocketConcurrencyTag, async () => {
-      await this.networkAvailableTask?.promise;
-      this.logger('debug', 'Restoring SFU WS connection');
-      this.cleanUpWebSocket();
-      await sleep(500);
-      this.createWebSocket();
-    }).catch((err) => this.logger('debug', `Can't restore WS connection`, err));
   };
 
   get isHealthy() {
@@ -409,7 +408,7 @@ export class StreamSfuClient {
     data: Omit<JoinRequest, 'sessionId' | 'token'>,
   ): Promise<JoinResponse> => {
     // wait for the signal web socket to be ready before sending "joinRequest"
-    await this.signalReady;
+    await this.signalReady();
     if (this.joinResponseTask.isResolved || this.joinResponseTask.isRejected) {
       // we need to lock the RPC requests until we receive a JoinResponse.
       // that's why we have this primitive lock mechanism.
@@ -478,7 +477,7 @@ export class StreamSfuClient {
   };
 
   private send = async (message: SfuRequest) => {
-    await this.signalReady; // wait for the signal ws to be open
+    await this.signalReady(); // wait for the signal ws to be open
     const msgJson = SfuRequest.toJson(message);
     if (this.signalWs.readyState !== WebSocket.OPEN) {
       this.logger('debug', 'Signal WS is not open. Skipping message', msgJson);